Add files via upload
So this is doing a lot of things. Right now it just transfer all the data from the model and experimental of a certain trial to a fancy excel document. Soon, it will also plot the data, and calculate some chi^2 and R^2 values. Or that might end up being a different file which it imports. IDK